gRPC proxy API
This changes the interaction model from the unary `invoke` method to
what is effectively a gRPC proxy that accepts any request and dispatches
it to the Oak Node.

Currently the dispatch is dynamic, and the raw gRPC method name and
input and output bytes are exposed to the Oak Node, but in the future we
will have a layer of code generation to make this more type safe.

This allows clients to use an unmodified gRPC client to interact with
the Oak Node, as long as it establishes the connection in a secure way
using EKEP.

Also turn the example directory into a single Rust workspace, so that
dependencies are shared between crates. See
https://doc.rust-lang.org/book/ch14-03-cargo-workspaces.html .

Note that some of the examples will have to be fixed after this change,
but I will do that in a separate CL.

Change-Id: I1ffe8d7c1fda9a1b13a05fe4edf98132f43d8d10