Add logic for configuring assertion authorities to OakManager
This change removes the logic in EnclaveServer for initializing Asylo
assertion authorities. As assertion authority libraries are global to
the enclave instance, they are initialized in the enclave in the setup
of TrustedApplication. The enclave's EnclaveConfig can be configured to
explicitly initialize each assertion authority that is used in the
enclave.

This change requires updating to Asylo v0.3.4.1, which is a hotfix
release that opens up visibility to a required dependency.