Make channels message-oriented and return status values
To guarantee that a read only ever produces a complete message, we need
to cope with the case where the provided buffer for channel_read is not
large enough.  In this case, no data is returned but the required size
needs to be, so the caller can re-allocate and call again.

Use a new actual_size parameter that gives a memory location to store
the relevant size, and shift the return value to be a status enum
(which includes a 'buffer too small' value).

Also shift channel_write to return a status enum too.

In turn, this means that the Rust code for a ReadChannelHalf cannot
implement the Read trait any more, because that trait is fundamentally
stream-oriented not message-oriented (i.e. it assumes partial reads are
just fine, and just repeats until EOF).