Define rust version to install in Dockerfile
The current dockerfile install the latest version of Rust at the time
the Docker image is built, which is obviously not reproducible.

In particular, there has just been a [new Rust
release](https://blog.rust-lang.org/2019/07/04/Rust-1.36.0.html) and our
tests are now suddenly failing because of some incompatible changes,
even though we have not modified anything in our code to trigger the
failure.

Also install components in the same command, and additionally install
Rust Language Server, so that the container can be used for Rust
development if desired.