Use local status.proto in BUILD files
The Rust code needs a local status.proto in order to be able to
generate status.rs, so use this local version in C++ code too,
rather than the @com_google_googleapis version.

This has the side effect of fixing:
    bazel build //oak/proto:storage_cc_proto

However, the grpc target (:storage_cc_grpc) is still not working.