Move Node event loop into Wasm
Replace the 3 required entrypoints (oak_initialize, oak_handle_grpc_call
and oak_finalize) with a single oak_main entrypoint that is not expected
to terminate.  Shift the C++ code to run this entrypoint in a new thread
of its own (and detach() it).  This also allows the removal of the
global Node instance in the Oak Rust library.

Add initial implementation of a wait_on_channels() host function that
allows the Node code to wait for pending inputs across channels,
although the initial implementation merely waits for something to arrive
on the single hard-coded gRPC-inbound channel read half.

Add helpers in the Oak Rust library that manage the buffer used to pass
channel wait information back and forth.

Add BlockingRead() and Await() methods to C++ Channel code, which block
until a message is available to read on the Channel.  Shift to use this
in both the new wait_on_channels() host function, and in the code that
retrieves the gRPC responses that result from a gRPC Node invocation.

Add a canonical event_loop(node) method in the Oak Rust library that
reads gRPC messages from the grpc-in channel and calls the node's
invoke() method, and invoke this from the oak_derive::OakExports macro.

This allows all of the existing Oak examples to be converted to the
Wasm-based event loop model with almost no changes to the code (just
need to add a `use oak::OakNode` to bring the trait into scope).