Block OakNode destruction on thread completion
Store the main thread in the OakNode class, and add a destructor
implementation that joins this thread before completing.

This means that attempts to delete OakNode instances will block, but
this is safer than the alternative (leaving the still-running thread
to reference deleted data).

The examples will still run forever, but the testdata/minimal.wat
test case does terminate.