Refactor the oak_runtime and the server implementations (#184)
These commits are refactoring the previous implementation of the oak runtime and fixes a few issues.

I moved the actual server creation from the runtime to the server class. This will allow us to create the server differently for each implementation. Right now this leads to some duplication, but as we will move the asylo server to support remote attestation, things will diverge.
Examples can now use insecure channels (for the dev version).
Remove most asylo dependencies from dev
Fixed an issue where I didn't created a server and a runtime for each "application" in the dev version.
Fixed a race condition where the queue was created before the server