Refactor server code outside of the manager (#203)
This is another step towards having a fully configurable set of channels
and nodes. It introduces a special (for now) new node that is the GRPC
one. It also decouples the manager-server ownership and moves it to the
runtime

This allows the wasm nodes to stop being gRPC servers.

Next, we need to create a base class for all nodes and start migrating
towards a model where all channels and nodes are configurable and owned
by the runtime