Example multi-node application
The multinode application has two nodes, a "frontend" and a "backend".
The frontend receives gRPC requests from the gRPC pseudo-Node, and
forwards the content to the backend.  The backend transmutes the content
and returns it to the frontend, which replies over gRPCs.

This example application deliberately does things a little differently
than other examples, so we can be re-assured that conventions/defaults
are not actually hard requirements.

 - The frontend and backend communicate over the raw channel without
   using gRPC/protobuf encodings (they use serde_json instead).
 - The application configuration uses non-default names for nodes and
   ports.
 - The backend node does not use the oak_derive crate, nor does it
   use the SDK's event_loop() helper.