Add 'expect' macros for use in ABI tests
Using assert_eq! would be more familiar, but it generates a panic
on failure which would need unwind protection to catch. (The tests
are methods on a FrontendNode type, which does not implement the
UnwindSafe trait.)

So clone assert_eq! into a simplified variant that returns an error,
and clone assert_matches! similarly.