Cache docker images to Google Container Registry (#250)
For now, we still rely on logged-in users to manually push a new version
of the image when dependencies change, in the future we should automate
that too. CI and other users can already rely on those images to avoid
rebuilding the images locally every time, since the Registry is set to
allow public access.