Use Google Cloud Build to upload artifacts
Add an additional pipeline running on Google Cloud Build, which builds
the enclave binary and pushes it to Google Cloud Storage. The intention
is for this to be used to build reference binary artifacts to be
deployed on Borg in Google if necessary, once we get full reproducible
builds (#241).

Always use Bazel cache for builds and for tests.

Example run:
https://pantheon.corp.google.com/cloud-build/builds/700d7aea-6799-4822-afa1-ec3758e88faf?project=oak-ci