Information flow initial tracking
Define secrecy and integrity labels for Nodes in the Application
configuration, each of which is a set of tags.  This becomes the
initial per-Node labels, which are then:
 - attached to outgoing Messages
 - updated according to incoming Messages.

The secrecy labels for a Node are also updated with the per-request
tags that arrive on incoming gRPC requests.

However, the labels/tags do not yet connect to anything: labels are
propagated and updated, but there is no label-specific functionality
in the Oak ABI and the labels have no effect on anything (except log
messages).