Convert storage API to gRPC
This is using gRPC to encapsulate request and response messages to and
from the storage node, but not actually using any transport from gRPC
itself.

Essentially it is a DSL that describes how to interact with the storage
node from other nodes, and as such it can rely on code generation to
create type-safe clients from other nodes, and also ensure that the node
implementation conforms to the same interface, rather than using
distinct and unrelated sets of request and response messages.