oak_tests: make test runtime a global static
Use lazy_static and RwLock to have a global mutable runtime to use
for tests.  This is necessary because the code under test expects to
use global TCB entrypoints (e.g. channel_write), so when these are
mocked out for tests the test infrastructure also has to be global.

This also means that tests that use the mock TCB need to be serialized,
which can be done (as here) with the serial_test crate.