abitest: add a unit test for example
This runs the ABI test application against the oak_tests test
framework, which helps to confirm that the latter behaves similarly
to the real TCB.  Skip tests that involve raw pointers though.

Add an abitest/tests Rust library which has test code to exercise
the abitest example in a multi-Node configuration.

Requires the following changes to the existing Rust code:
 - Allow the modules to be built as `lib` as well as `cdylib`.
 - Only define the global `oak_main` entrypoint if compiling for Wasm,
   so that it is possible to link in both the frontend and backend
   nodes without getting a duplicate symbol.
 - Immediately delegate from each `oak_main` instance to a public
   `crate::main` function; these can be used as the per-Node
   entrypoints for testing.  Keep the unwind-protection in just
   `oak_main` though, as it's only needed for the FFI boundary
   there (and it's also useful for panics to be emitted during
   testing).
 - Export the generated code for the protobuf message definitions.
 - Ignore failures from oak_log::init().