Support linking Rust code in Oak Runtime
Currently only compiled with `no_std`, therefore only stack allocation
is allowed. In the future we should plug in a real allocator and link to
`liballoc` to make it more useful.

Note to reviewers: I have added a simple check in `oak_runtime.cc` to
verify that the bindings are working, but let me know if you prefer to
leave that out.

Ref #313