examples/chat: use a new Node for each room
Split out a backend Node type that just echoes anything sent
to it, and use this to perform per-room function.

On starting a new room, create a new instance of this Node.
Whenever anyone joins the room, pass the response writer channel
handle to the room to add to its list of places to fan out to.

Move the app's proto definition into a shared common crate.