scripts: allow for extra dependencies for examples
If the file ./examples/<name>/run_deps is present, it should hold
a command that needs to be run in the background before the example
is run (and killed afterwards).

For the hello_world example, run a storage provider on the default
address while the example is running, so the storage interactions
get exercised a bit.