Rename chat example modules
The `--module` flag was assuming that the modules are specified as
"backend,frontend", but the run_example script was collecting them in a
different order, which ended up mixing up the two modules needed for the
chat example.

This commit renames the modules so that it's clear in which order they
are going to be specified, and flips the expected order of the module
flag to reflect the fact that the frontend node is the first one to be
specified.