runtime: specify entrypoint name dynamically
Change the node_create() ABI function to also take an entrypoint name.
Re-organize the ApplicationConfiguration protobuf to include the name
of the initial entrypoint in the initial Node; any subsequent Node
creation operations need to specify an entrypoint name that exists in
the appropriate Wasm Node configuration as an exported function of
the correct signature (`fn(u64) -> ()`).