Tweak build_gh_pages script (#582)
Remove diff logic so that it can be run at any commit and just push the
latest version, even if it causes extra diffs because of the
non-determinism. Incidentally, it seems to be deterministic on my
machine (but maybe it's just because generated docs are cached).

Add generation of index.html file.

Use a temporary directory to stage the output of the generation process.