Support structured output in runner script
The downside is that it now requires all the steps to have finished
before printing anything at all, which makes it hard to figure out at
what point of the overall execution it is.

On the other hand this will make it easier to support future-based
return value in the future, at which point we will reimplement the
streaming-line output functionality.