runtime: use Notification object to avoid polling (#589)
Each Node thread that is doing a wait_on_channels() blocking host
function call creates an Notification object, and passes a weak_ptr
to this when checking the channel's read status.

If the channel is waitable (empty, not orphaned) then this weak_ptr
to the Notification object is added to the channel's pending-list.

Repeat this for each channel included in the wait operation.

When a write to the channel occurs, all of the Notification objects
in the channel's pending-list are Notify-ed, provided that the held
weak_ptr can be converted to a shared_ptr (i.e. that some notification
from a different channel hasn't already triggered the Notification and
ended the wait_on_channels).

The Notification object uses the build-in condition variable of an
absl::Mutex to signal that a Notification is ready.  Allow multiple
triggers of the Notify() method (unlike absl::Notification).