Introduce label parameter to `Runtime::node_create`
This is not part of the Wasm ABI yet, just part of the implementation of
the Rust oak_runtime, but not used for information flow yet.

Unify Label trait and struct in oak_abi (there is no strong reason for
the separation at this point, though we may want to reintroduce it in
the future).

Rename `Label::can_flow_to` to `Label::flows_to` in accordance with
literature.

Add various missing doc links.

Ref #603