storage: re-use oak::grpc helpers
Change the storage interface so that each storage request is expected
to be in a channel of its own, rather than flowing on the original
channel.

This makes the storage component's use of channels more directly
equivalent to how the gRPC server and client pseudo-Nodes use
channels.

In turn, this allows for various storage-specific bits of code to be
removed and replaced by use of oak::grpc helper functions.

Along the way, commonize variable naming between the storage code and
the gRPC client pseudo-Node code.