runtime: pass initial handles around
At start of day there are exactly two Nodes that need to be started,
the gRPC server pseudo-Node and the initial Application Wasm Node.
Hang on the OakNode reference and initial Handle for both of these,
and move OakNode::Start() to always take a Handle.  This means that
OakNode::SingleHandle() can be removed.