examples: demonstrate Go client
Move the client for the translator example to use Go rather than
C++ to demonstrate language interoperability.

Update WORKSPACE to pull in Bazel Go rules, Gazelle, and repos
for the dependencies of the Go code. For now, manually add
go_repository() clauses for the (transitive) dependencies; if more
Go code is added in future, then this will best be handled by
Gazelle.

Include gen.go and generated translator.pb.go, which allows the
client to be built with `go build ./translator.go`.

Also update the Programming Oak doc to mention ability to use clients
in different languages.