runtime/rust: move invalid handle handling out to wasm code
The methods of the Runtime that deal with sets of handles shouldn't need to
cope with Option<Handle> just to make a single user of the API easier.
Shift to a simpler API that reports on exactly the handles it's given, and
move the tracking of invalid handles out to the Wasm specific code.