runtime: rename runtime::Handle to ChannelHalfId
To avoid confusion, reserve the word "handle" for ABI Handles.  This
type identifies one half of a channel (either read or write), so give
it the name "ChannelHalfId". It also doesn't contain an ABI Handle,
so use raw u64 for its internals to avoid confusion.

Also avoid using "reference" for things that aren't Rust references.

The contents can also move back to being private, as
pseudo-Nodes now use ABI handles throughout (which is a partial revert of
commit 93a4b30eecae, "runtime/rust: make handle/ID structs public").