runtime: use Arc<Channel> in ChannelHalf
Rather than using a new numbering space of u64 values to indicate a
reference to a channel, stick to the standard Arc<Channel> type.

Use the counts in a Channel to refer to the number of ChannelHalf
objects in existence, each of which holds an Arc<Channel>.

The actual number of Arc<Channel> references will be one higher, due
to the reference stored in ChannelMapping.  In ChannelMapping:drop_half,
watch out for these counts reaching zero, and remove this final reference
(which should then drop the Channel).

Note that cycles in Channel <-> Message references will lead to leaks.