runtime: commonize Node creation
Commonize all of the boilerplate for starting a thread for a particular kind
of Node. This also means that code for the Nodes doesn't need to know about
threading, and the cleanup of resources held by Nodes is under the control
of the Runtime, rather than the code for each Node. These is also now exactly
one place that removes the state for a Node.

Use a consistent node name everywhere.