runtime: move proxy and handle to run()
Make all of the Node types more functional-style by passing the
RuntimeProxy and initial Handle to the Node's run() method, rather
than including this information in the constructor and having each
Node type store it in its struct, only to be used later in run().

The only Node type that needed tweaking for this is the Wasm Node,
which was using the RuntimeProxy at create time in its checks of
Wasm entrypoint validity.  To cope with this, make the runtime field
of WasmInterface an Option, so the entrypoint-checking version will
have None (which makes more sense: there shouldn't be any possibility
of entrypoint-checking actually invoking a Runtime entrypoint).