Take into account downgrade privilege when reading
Previously downgrade was only allowed when writing to lower labels, but
not when reading from higher labels. This makes both possible, relying
on the same underlying Node privilege.

Credit to @aferr for the explanation and suggestions on how to implement
it, and also link to the relevant Slack discussion, for reference:
https://project-oak.slack.com/messages/CHE9E13C3/p1590712037024800?thread_ts=1590660390.016600

Ref #630