Propagate gRPC request label
- Enforce that **exactly** one label is specified per gRPC request. From
now on, the gRPC server will reject requests with no label attached, or
with multiple labels. It is possible to attach a "public" label, but it
needs to be specified explicitly.

- Change the examples to attach such "public" label, until we figure out
the rest of the downgrading functionality in the Oak Runtime.

- Change return type of `handle_grpc_request` and `handle_grpc_response`
from `Result<GrpcResponse, OakStatus>` to `Result<GrpcResponse, ()>`, to
make it clear that all errors are actually logged within the body of the
functions. Previously there was a comment somewhere else that was
explaining why those errors were being ignored, and also a statement
that confirmed that they were all logged (which turned out to
be false).

- Log errors generated by gRPC channel creation.

- More uniform error handling in C++ example clients.