Improve logging of gRPC request handling
The main change is that we log one info entry for each gRPC request,
including details of its label, so that it's possible to debug label
mismatch issues.

Also do not log the contents of the request, only its size, and the
client IP address (though this does not seem to work reliably yet).