runtime: re-use gRPC client connection
As long as there is no error encountered when processing a method invocation,
store the client object for re-use on the next method.

However, if an error is encountered in the gRPC processing itself, drop the
client object so the next method will re-connect.  Note that a non-OK status
code from the remote server doesn't count as an error in this respect.

Fixes #1057.