abitest: don't re-use gRPC client pseudo-Node
Each run of the abitest client starts a new gRPC server that happens to re-use
the same port as on the previous run.

Create a gRPC client pseudo-Node for each test case rather than re-using
a Node-global version to allow for this.

Without this, a second run of the abitest client will use the same
gRPC pseudo-Node as the previous run, and the pseudo-Node will attempt to
re-use its connection but the process it was connected to has gone.