runtime: restructure gRPC server processing
In preparation for adding server-streaming method support in future,
reorganize the invocation processing to put all per-response code in
a (single-shot) loop.

Convert read_grpc_response() into an Iterator to feed this loop.

Also, put the closing of the response channel into a manual Drop
implementation for the Iterator, to ensure it always gets done
automatically.