examples: move tests to be integration tests (#1154)
Most of the tests for example Oak Applications just use the external
gRPC service definition for testing, and so can be external integration
tests rather than unit tests.

This does require an "rlib" target for the modules though, because of
https://github.com/rust-lang/cargo/issues/6659

The trusted_information_retrieval example's tests included both unit
tests and integration tests, so is split.  Along the way, make the
floating comparisons in the unit tests a bit more robust.

Fixes #1054.