Only allow public Nodes to create entities
Implement `Clone` and `PartialEq` for `NodeMessage` to facilitate
testing.

Add termination check to `channel_create` host function.

Add more logging around labels in various host functions.

The tests are not as comprehensive as I would like them to be, since it
is hard to simulate the case in which a public node creates a non-public
channel and passes that to a non-public node.

Fixes #1143