runtime: don’t expose internals of Runtime
The initial implicit Node is only used for injection of the start-of-day
ConfigMap, so move that functionality inside configure_and_run().

This means there is no longer a need to make the `RuntimeProxy` (and
with it the `NodeId` type and various other modules) visible to the
world outside the `Runtime`.

Also has knock-on effects in the oak_tests crate, and on all the
integration tests that use oak_tests.