Grant TLS tag privilege to gRPC client pseudo-node
Replace the certificate SAN name field in the label, with a field that
represents the authority (host:port) of the remote endpoint instead.

Make `run_event_loop` more type safe by passing in a `Reciever<T>` where
`T` is the same type of the `Node<T>` that actually handles the
messages.

Create an explicit `GrpcServerInit` message to pass to a gRPC server
pseudo-node, instead of manually passing handles to it in an
unstructured way.

Remove unused `address` field from `GrpcClientConfiguration` protobuf
message.

Split aggregator example in two nodes, so that the gRPC worker node may
be assigned a different label in the future, to reflect the fact that it
is handling non-public data.

Add explicit scheme to gRPC client URL in aggregator example (not sure
how / whether it worked previously, Tonic complains).

Make aggregator client submit a meaningful number of samples in order to
reach the aggregation threshold, instead of just one.

Ref #814