Support "spread" operator via macro (#1237)
This allows removing a layer of nesting in various places where we used
to use `flatten` instead.

A common case is for lists of flags, which sometimes contain other lists
of flags, which need to be expanded in line.

The implementation is not very efficient and has some corner cases (e.g.
still requires calls to `to_string` to make things work sometimes).