Proof of single-step noninterference with one call
This is a proof of a single step of noninterference without downgrades,
sometimes referred to as "the unwinding condition" in a formal
verification context, or "single command noninterference" in a
programming languages context. This means that if we run the state
transition function for just one step, then noninterference is enforced for
that one step. At the moment, the model just has a single call that is
actually properly modeled (the write call), so this is not really a complete
proof. The definition of noninterference is also still undergoing revision.
The proof itself is also not terribly automated and almost no effort was
put into making it elegent yet.