Formalize concurrent execution /w two-tier semantics
The concurrent execution of nodes is now modeled using a two-tiered
semantics. There are two sets of (related) inference rules: a thread-local
semantics and a global semantics. The thread-local semantics defines
execution of a command by a particular node. The global semantics
non-deterministically selects which node should execute, and updates the
state by using the thread-local semantics with the chosen node. This
two-tier semantics is important for the noninterference proof because it
gives us a way to model the fact that nodes can execute with an
arbitrary interleaving, and we want to show that there are no
information leaks regardless of scheduling (aside from perhaps
secheduling side channels).

The thread-local semantics closely resembles that of an imperative
programming language. A key difference though is that in order to be
language-agnostic (we cannot model the code that actually runs inside a
node aside from execution of the ABI calls) the semantics does not
contain usual constructs for control flow (like sequential composition
or loops). Instead, after any command is executed, a new command is
chosen nondeterministically (though the update to this command happens
in the global semantics). This models the fact that a node can
execute arbitrary code aside from the ABI calls in between calls to the
ABI. The choice of next command to execute is nondeterministic because
the next command does change, but there are no premises about the new
command in the inference rule.

The global semantics _nondeterministically_ chooses the thread to execute
because there are no premises about the thread that is used in the
thread-local semantics.

This two-tier semantics is inspired by "Secure Information Flow in a
Multi-threaded Imperative Langiage" by Smith and Volpano '98
(http://web.cs.ucla.edu/~palsberg/tba/papers/smith-volpano-popl98.pdf).
This same approach of modeling concurrent execution using a two-tier
semantics is also taken in "Compositional Non-interference for
Concurrent Programs via Separation and Framing."