Restricted unwinding proof with concurrent model
The nondeterminstic semantics breaks the previous proof of single-step
noninterference because now we need to reason about tricky edge cases in
scheduling behavior: for example, a "high" node might execute different
commands in each of the two traces. (One standard way of dealing with
that case is to prove "clearance": that high commands do not affect low
state). This commit just proves single-step noninterference for the case
where the chosen node performs the same command in both traces.