Event-augmented semantics
This commit introduces "events" that model inputs and outputs to nodes
(for example, during the channel_read and channel_write calls). At the
moment an event is a label and either: an input message, an output
message, or nothing. In the future, events will also be used for modeling
values that are downgraded.

Events are needed for specifying noninterference in this context. An
alternative way of writing noninterference (without events) would be to just
use relations on states. This is not sufficient for the oak runtime model,
because we cannot model all the state in a node, and so a definition
that uses states alone and not events would not have any way of tracking
the read calls.

Events are introduced into traces of system execution by using an
augmented semantics. The augmented semantics works just like the normal
one except that it adds these events to the trace. The idea is that
noninterference proofs will be done on the augmented semantics and there
will also be proof that the augmented and "normal" semantics are equal.
(This equality proof should be trivial).