EvAug cuts out states; progress on PossibilisticNI
The EvAug semantics is refactored so that it no longer is a state ->
trace -> state -> trace -> Prop relation but instead just a trace ->
trace -> Prop relation. This removes the need to have various
propositions about the states and traces involved in the step relation
agreeing. This also updates the statement of possibilistic NI and the
single step theorem. In the process, some problems with both of those
theorems are fixed. There is also some progress on the proof
for possibilistic NI. It is mostly done aside from some smaller lemmas
that I think are true but are admitted (the proof of the single-step
theorem is also still admitted).