Use Docker BuildKit for the CI and docker_script
Docker BuildKit has a number of advanages over the tradiditonal
Docker back end - it's potentially quicker as it can make use of
parallel jobs when multi-stage Dockerfiles are used, and it also
enables the use of automatically set target/host ARGs, such as
TARGETARCH. This is especially useful when working with multiple
target architectures.

This change modifies the main docker image name to indicate it is
using BuildKit. This is a WIP testing change.