Fix channel_write channel_read calls.
In particular, the model now includes the options to send and receive
sets of channels (in addition to messages) as a part of the channel_send
and channel_receive calls.

Because read handles are expected to be linear, all the sent handles are
deleted from the sender during a channel_send.