Possibilistic single-step for system step
This commit includes a proof of possibilistic_ni_1step in
PossibilisticNI.v, which is the unwinding condition. However, it still
depends on several theorems are still admitted, but should be true. The
definition for havocking calls for nodes that have just executed has
also been fixed in the system-wide state transition relation.
Previously, the call havocking part of the definition assumed that the
contents of the node have not changed as a result of the node's
execution, but clearly that is wrong. Now it properly updates the
(possibly changed) node that is pointed to by the node id that was
executed.