Make run_ci more platform-agnostic.
This change modifies the default rust targets that are used
when no overrides are specified, so that the assumption is that
run_ci (and other commands) default to the current platform.
In the case of 'macos' this requires a fudge to chenge the Rust
target triple to include 'darwin'.

The default are now $ARCH-unknonwn-$OS-musl for the server,
and $ARCH-unknown-$OS-gnu for the clients.

This does not affect the existing overrides available for the other
runner commands.

This change makes a number of modifications to the build scripts
and the arm64v8 Dockerfile in order to support running the CI
for Oak on the Packet eMAG systems.

Summary:
 - The Dockerfiles can be of the form $ARCH.Dockerfile
 - Images are created of the form oak-$ARCH for the main
   build and the oak servers.
 - x86_64 should run unmodified
 - Support for running CI in docker locally by overriding
   DOCKER_REPO to ""
 - Image names and locations can be overridden.

The intent here is to make the build system as generic as possible
without becoming over-complex: There is a single 'hack' to the
scripts/common file that is required to account for the differing
build stages on aarch64 (build Bazel from source).

Image management:
This means that a pre-prepared image for bazel-bootstrap-arm64v8
and bazel-build-arm64v8 should be pushed to the repo before
further integration, as otherwise the builds will be very slow.