Fix oak_loader being compiled on each runner run
Background: The introspection client is build on each run, changing the timestamp on output assets, even if the resulting file is identical. Cargo checks this timestamp to determine if these files have changed. Since timestamps are different on each run, cargo recompiles each binary that includes these files, on every run. This patches this via a webpack plugin that ensures that the timestamp of the introspection assets only changes if the contents changed.