Write call single-step, single-thread unwind
This commit finishes a proof of the single-step, single-threaded
unwinding condition for just the write call and just in the case where
the label of the node making the call does flow to the label of the
observer. Several new smaller theorems are introduced but admitted for
now. They should be relatively simple. The files are reorganized a bit.