Move back to GenericMap
Following discussion with @jadephilipoom, this commit removes the ssreflect
finmap in favor of using the old simpler GenericMap.v implementation of
finite maps. SSReflect is a fairly complicated library that turns out to
be closer to a whole new dialect of coq. Given that the only use for it
was finite maps and we are not doing computation on maps, it seems
best to just use simpler maps.