Move tactics into their own file (#1510)
In preparation for more proof automation, this commit moves the more general-purpose tactics from TraceTheorems.v into their own file. Since the tactics will now be exported for use in other files, it also adds tests to ensure that the behavior of the tactics isn't accidentally changed in unexpected ways.